Release areca-7.3.2

+++ src/com/myJava/file/delta/tools/IOHelper.java
@@ -89,7 +89,7 @@ public class IOHelper {
     		return 0;
     	}
 
-    	int nb = 0;
+    	long nb = 0;
     	while(nb < length && in.read() != -1) {
     		nb++;
     	}
